# improve performance of normalizeList