--- a/lib/normalizeList.js
+++ b/lib/normalizeList.js
@@ -9,15 +9,20 @@
  */
 
 function normalizeList (val, name) {
-  const list = [].concat(val || [])
-
-  for (let i = 0; i < list.length; i++) {
-    if (typeof list[i] !== 'string') {
-      throw new TypeError(name + ' must be array of strings or false')
+  if (typeof val === 'string') {
+    return [val]
+  } else if (val === false) {
+    return []
+  } else if (Array.isArray(val)) {
+    for (let i = 0, il = val.length; i < il; ++i) {
+      if (typeof val[i] !== 'string') {
+        throw new TypeError(name + ' must be array of strings or false')
+      }
     }
+    return val
+  } else {
+    throw new TypeError(name + ' must be array of strings or false')
   }
-
-  return list
 }
 
 module.exports.normalizeList = normalizeList